% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeqd.buffer.R
\name{aeqd.buffer}
\alias{aeqd.buffer}
\title{Buffer latlon polygons in metres}
\usage{
aeqd.buffer(poly, width)
}
\arguments{
\item{poly}{TODO}

\item{width}{TODO}
}
\description{
TODO
}
\note{
Modified from https://gis.stackexchange.com/questions/121489/1km-circles-around-lat-long-points-in-many-places-in-world

License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
