% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.yieldpoints.gui.R
\name{filter.yieldpoints.gui}
\alias{filter.yieldpoints.gui}
\title{GUI for filtering yield data}
\usage{
filter.yieldpoints.gui(indata, inlayer, samplesize)
}
\arguments{
\item{indata}{TODO}

\item{inlayer}{TODO #TODO maske optional}

\item{samplesize}{TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
