% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.yieldpoints.R
\name{filter.yieldpoints}
\alias{filter.yieldpoints}
\alias{filter.yieldpoints.reset}
\alias{filter.yieldpoints.reset}
\alias{filter.yieldpoints.resample}
\title{Update filter values (altering the original dataframe)}
\usage{
filter.yieldpoints(indata, metric, value, inlayer = NA, byfield = FALSE,
  samplesize = 1e+05)

filter.yieldpoints.reset(indata, filters = NA)

filter.yieldpoints.resample(indata)
}
\arguments{
\item{indata}{TODO}

\item{metric}{one between f_rangev, f_smv, f_mins, f_rangey, f_stdy, f_rangeq, f_pos}

\item{value}{value or values to apply (depending on metric)}

\item{inlayer}{optional: SpatialPolygonsDataFrame of fields (a field "idfield" must be present if byfield is TRUE)}

\item{byfield}{if FALSE (default), consider data as a unique field; if TRUE, iterate each filter on field "idfield"}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 100000; if NA: all the points)}

\item{filters}{TODO}

\item{indata}{TODO}

\item{indata}{TODO}
}
\description{
TODO

TODO

TODO
}
\note{
License: GPL 3.0

License: GPL 3.0

License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}

Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}

Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
