% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.yieldvgm.gui.R
\name{fit.yieldvgm.gui}
\alias{fit.yieldvgm.gui}
\title{GUI for filtering yield data}
\usage{
fit.yieldvgm.gui(indata)
}
\arguments{
\item{indata}{TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
