% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_yieldvgm_gui.R
\name{fit_yieldvgm_gui}
\alias{fit_yieldvgm_gui}
\title{GUI for filtering yield data}
\usage{
fit_yieldvgm_gui(indata)
}
\arguments{
\item{indata}{TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
