% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.yielddata.gui.R
\name{import.yielddata.gui}
\alias{import.yielddata.gui}
\title{GUI for loading raw yield data}
\usage{
import.yielddata.gui()
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
