% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.yieldmaps.R
\name{merge.yieldmaps}
\alias{merge.yieldmaps}
\title{Mosaic field tiffs}
\usage{
\method{merge}{yieldmaps}(indir, reg_ex = "\\\\.tif$", outfile)
}
\arguments{
\item{indir}{the directory in which field maps are stored (equivalent of outdir+subdir in interp.yieldpoints())}

\item{reg_ex}{\code{character (1)} optional regular expression to be applied of indir content}

\item{outfile}{output raster file (with path)}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
