% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.yielddata.R
\name{read.yielddata}
\alias{read.yielddata}
\title{Read yield data}
\usage{
read.yielddata(input, format)
}
\arguments{
\item{input}{TODO}

\item{format}{\code{character (1)} TODO}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
