% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_input_shapes.R
\name{shiny_input_shapes}
\alias{shiny_input_shapes}
\title{Load shapefiles from UI}
\usage{
shiny_input_shapes(input_paths, multiple = TRUE)
}
\description{
internal TODO
}
\section{Slots}{

\describe{
\item{\code{input_paths}}{TODO}

\item{\code{multiple}}{TODO}
}}

\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
