% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.data.table.R
\docType{class}
\name{yield.data.table}
\alias{yield.data.table}
\alias{yield.data.table}
\title{Class yield.data.table}
\usage{
yield.data.table(datatable, comm_humid = as.numeric(NA),
  max_width = as.numeric(NA), time_step = as.numeric(NA),
  crs = CRS(as.character(NA)), format = as.character(NA))
}
\arguments{
\item{datatable}{\code{data.table} TODO}

\item{comm_humid}{\code{numeric (1)} TODO}

\item{max_width}{\code{numeric (1)} TODO}

\item{time_step}{\code{numeric (1)} TODO}

\item{crs}{\code{CRS} TODO}

\item{format}{\code{character (1)} TODO}
}
\description{
TODO

TODO
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{data.table} TODO}

\item{\code{comm_humid}}{\code{numeric (1)} TODO}

\item{\code{max_width}}{\code{numeric (1)} TODO}

\item{\code{time_step}}{\code{numeric (1)} TODO}

\item{\code{crs}}{\code{CRS} TODO}

\item{\code{format}}{\code{character (1)} TODO}
}}

\note{
License: GPL 3.0

License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}

Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
