% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.makegrid.R
\name{yield.makegrid}
\alias{yield.makegrid}
\title{Create global output grid}
\usage{
yield.makegrid(indata_sp, outres = 5, border = 15,
  outcrs = indata_sp@proj4string)
}
\arguments{
\item{indata_sp}{TODO all input points}

\item{outres}{\code{numeric (1)} TODO spatial resolution of output raster (in crs unit)}

\item{border}{\code{numeric (1)} TODO extent of border (in crs units)}

\item{outcrs}{\code{CRS} output CRS}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
