% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.makespdf.R
\name{yield.makespdf}
\alias{yield.makespdf}
\title{Make SPDF from filtered yield data}
\usage{
yield.makespdf(indata, outcrs = CRS("+init=epsg:3035"))
}
\arguments{
\item{indata}{TODO}

\item{outcrs}{\code{CRS} output CRS}
}
\description{
TODO
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
