% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldmaps_process.R
\name{yieldmaps_process}
\alias{yieldmaps_process}
\title{Process a group of fields}
\usage{
yieldmaps_process(rese_input, inlayer, grid_path = NA, filtered = TRUE,
  id_fieldname = "idfield", interp_dir = tempfile(), out_crs = NA,
  samplesize = 10000, buffer_radius = 15, vgm = NA)
}
\arguments{
\item{rese_input}{TODO}

\item{inlayer}{SpatialPolygonsDataFrame of fields (see also \code{id_fieldname})}

\item{grid_path}{optional: TODO}

\item{filtered}{logical: is \code{rese_input} already filtered? If TRUE (default),
no additional filter is applied; if FALSE, a standard automatic filter is applied.}

\item{id_fieldname}{optional: name of the \code{inlayer} field containing unique ID of fields (default: "idfield")}

\item{interp_dir}{directory where rasters are stores (default: temporary directory)}

\item{out_crs}{optional: CRS of output raster (default: CRS of rese_input)}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 10000; if NA: all the points)}

\item{buffer_radius}{numeric: buffer (default: 15) to be applied internally to field borders.}

\item{vgm}{named list of variograms (names must correspond to values of field \code{id_fieldname}).
If NA (default), they are automatically computed.}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 100000; if NA: all the points)}
}
\description{
This function calls other functions of the library to perform
interpolation over different fields.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
