% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.yieldpoints.R
\name{filter.yieldpoints}
\alias{filter.yieldpoints}
\alias{filter.yieldpoints.reset}
\alias{filter.yieldpoints.resample}
\title{Update filter values}
\usage{
filter.yieldpoints(indata, metric, value, inlayer = NA,
  id_fieldname = "idfield", byfield = FALSE, samplesize = 1e+05)

filter.yieldpoints.reset(indata, filters = NA)

filter.yieldpoints.resample(indata)
}
\arguments{
\item{indata}{TODO}

\item{metric}{one between f_rangev, f_smv, f_mins, f_rangey, f_stdy, f_rangeq, f_pos}

\item{value}{value or values to apply (depending on metric)}

\item{inlayer}{optional: SpatialPolygonsDataFrame of fields (see also \code{id_fieldname})}

\item{id_fieldname}{optional: name of the \code{inlayer} field containing unique ID of fields (default: "idfield")}

\item{byfield}{if FALSE (default), consider data as a unique field; if TRUE, iterate each filter on field `id_fieldname``}

\item{samplesize}{maximum size of the sample of the original data to work with (default: 100000; if NA: all the points)}

\item{filters}{TODO}
}
\description{
\code{filter.yieldpoints} updates filter values (altering the original dataframe).

\code{filter.yieldpoints.reset} resets all filter values to FALSE (altering the original dataframe).

\code{filter.yieldpoints.resample} resets sampling order (altering the original dataframe).
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2018) \email{ranghetti.l@irea.cnr.it}
}
